+++ main.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/itzmeanjan/harmony/app/bootup"
-	"github.com/itzmeanjan/harmony/app/txpool"
+	"github.com/itzmeanjan/harmony/app/mempool"
 )
 
 func main() {
@@ -93,7 +93,7 @@ func main() {
 
 	}()
 
-	go txpool.PollTxPoolContent(ctx, resources, comm)
+	go mempool.PollTxPoolContent(ctx, resources, comm)
 
 	// This is just a fancy of blocking execution, so that
 	// main go routine doesn't die & program keeps running
